requirements.txt
flask
flask-cors

# PyTorch CPU (latest stable)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.7.1+cpu
torchvision==0.22.1+cpu

numpy
opencv-python-headless
Pillow
gdown
gradio>=3.41
timm

# MobileSAM
git+https://github.com/ChaoningZhang/MobileSAM.git
